--- src/app/dashboard/page.module.css
@@ -1,115 +1,0 @@
-.container {
-  font-family: Arial, sans-serif;
-  margin: 0;
-  padding: 0;
-}
-
-.header {
-  background-color: antiquewhite;
-  color: #333;
-  padding: 20px;
-  text-align: center;
-}
-
-.main {
-  padding: 20px;
-}
-
-.form {
-  margin-bottom: 20px;
-}
-
-.text-input {
-  padding: 10px;
-  width: calc(100% - 80px);
-  margin-right: 10px;
-  border: 1px solid rgba(0, 0, 0, 0.1);
-  border-radius: 10px;
-  height: 40px;
-  outline: none;
-}
-
-.submit-button {
-  padding: 10px;
-  height: 40px;
-  width: 200px;
-  border: 1px solid rgba(0, 0, 0, 0.1);
-  border-radius: 10px;
-  border: none;
-  cursor: pointer;
-}
-
-.back-button {
-  padding: 10px;
-  height: 40px;
-  width: 200px;
-  border: 1px solid rgba(0, 0, 0, 0.1);
-  border-radius: 10px;
-  border: none;
-  cursor: pointer;
-  background-color: #fd0808;
-  text-decoration: none;
-  margin: 0 50px;
-}
-
-.back-button a {
-  text-decoration: none;
-  color: #ffffff;
-}
-
-.submit-button:hover {
-  background-color: antiquewhite;
-}
-
-.task-list {
-  list-style-type: none;
-  padding: 0;
-}
-
-.task-item {
-  padding: 10px;
-  background-color: #f4f4f4;
-  margin-bottom: 5px;
-  border-radius: 5px;
-  display: flex; /* Aplica o modelo de layout flexível */
-  justify-content: space-between; /* Distribui os itens uniformemente ao longo do eixo principal */
-  align-items: center; /* Centraliza os itens verticalmente */
-  padding: 10px; /* Adiciona um espaçamento interno */
-}
-
-.task-item.completed span {
-  text-decoration: line-through;
-  opacity: 0.5;
-}
-
-.div-centralizada {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 30px 0;
-}
-
-.task-button {
-  background-color: #fd0808;
-  color: #ffffff;
-  border: none;
-  padding: 5px 10px;
-  border-radius: 5px;
-  cursor: pointer;
-  margin: 10px; /* Ajuste o espaçamento entre o botão e o título da tarefa conforme necessário */
-}
-
-.edit-button {
-  background-color: #180779;
-  color: #ffffff;
-  border: none;
-  padding: 5px 10px;
-  border-radius: 5px;
-  cursor: pointer;
-  margin: 10px; /* Ajuste o espaçamento entre o botão e o título da tarefa conforme necessário */
-}
-
-.edit-button a {
-  text-decoration: none;
-  color: #ffffff
-}
