+++ src/components/dashboard/FormCreateTask.module.css
@@ -1,0 +1,35 @@
+.form {
+  text-align: center;
+  margin: 30px 0;
+}
+
+.input {
+  width: 80%;
+  padding: 10px;
+  border: 1px solid rgba(0, 0, 0, 0.1);
+  border-radius: 10px;
+  height: 40px;
+  outline: none;
+  margin: 20px 0;
+}
+
+.submit {
+  padding: 10px;
+  height: 40px;
+  width: 200px;
+  border: 1px solid rgba(0, 0, 0, 0.1);
+  border-radius: 10px;
+  border: none;
+  cursor: pointer;
+}
+
+.submit:hover {
+  background-color: #495464;
+  color: #F4F4F2;
+}
+
+@media (max-width: 650px) {
+  .input {
+    width: 90%;
+  }
+}
